app file restructure
