+++ src/layouts/DefaultLayout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom';
-import Footer from '../components/Footer';
-import { NavBar } from '../components/NavBar';
+import Footer from '../components/common/Footer';
+import { NavBar } from '../components/common/NavBar';
 import { Box } from '@mui/material';
 
 export default function DefaultLayout() {
