+++ src/main.tsx
@@ -4,7 +4,7 @@ import { ThemeProvider, CssBaseline } from '@mui/material';
 import { AuthProvider } from './context/AuthContext';
 import './index.css';
 import App from './App';
-import { theme } from './theme/defaultTheme';
+import { theme } from './theme';
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
