+++ src/routes/ProtectedRoute.tsx
@@ -1,26 +1,3 @@
-// import { Navigate, Outlet, useLocation } from 'react-router-dom';
-// import { useAuthContext } from '../context/AuthContext';
-
-// interface ProtectedRouteProps {
-//   allowedRoles?: string[];
-// }
-
-// export default function ProtectedRoute({ allowedRoles }: ProtectedRouteProps) {
-//   const { user, loading } = useAuthContext();
-//   const location = useLocation();
-
-//   if (loading) return <div>Loading...</div>;
-//   // Not logged in
-//   if (!user) {
-//     return <Navigate to="/" state={{ from: location }} replace />;
-//   }
-//   // Logged in but role not allowed
-//   if (allowedRoles && !user.roles.some((role) => allowedRoles.includes(role))) {
-//     return <Navigate to="/" replace />;
-//   }
-
-//   return <Outlet />;
-// }
 import { Navigate } from 'react-router-dom';
 import { useAuthContext } from '../context/AuthContext';
