+++ src/theme/typography.ts
@@ -1,0 +1,72 @@
+export const typography = {
+  fontFamily: 'Roboto, Inter, sans-serif',
+
+  // Headings
+  h1: {
+    fontSize: 'clamp(2rem, 5vw + 1rem, 3.5rem)',
+    fontWeight: 600,
+    lineHeight: 1.2,
+    letterSpacing: -0.5,
+  },
+  h2: {
+    fontSize: 'clamp(1.75rem, 4.5vw + 1rem, 3rem)',
+    fontWeight: 600,
+    lineHeight: 1.25,
+    letterSpacing: -0.5,
+  },
+  h3: {
+    fontSize: 'clamp(1.5rem, 4vw + 0.75rem, 2.5rem)',
+    fontWeight: 600,
+    lineHeight: 1.3,
+    letterSpacing: -0.5,
+  },
+  h4: {
+    fontSize: 'clamp(1.25rem, 3.5vw + 0.75rem, 2.25rem)',
+    fontWeight: 600,
+    lineHeight: 1.35,
+    letterSpacing: -0.5,
+  },
+  h5: {
+    fontSize: 'clamp(1.125rem, 3vw + 0.5rem, 1.75rem)',
+    fontWeight: 600,
+    lineHeight: 1.4,
+    letterSpacing: -0.5,
+  },
+  h6: {
+    fontSize: 'clamp(1rem, 2.5vw + 0.5rem, 1.25rem)',
+    fontWeight: 600,
+    lineHeight: 1.4,
+    letterSpacing: -0.5,
+  },
+
+  // Subtitles
+  subtitle1: {
+    fontSize: 'clamp(1rem, 2vw + 0.5rem, 1.25rem)',
+    fontWeight: 500,
+    lineHeight: 1.4,
+  },
+  subtitle2: {
+    fontSize: 'clamp(0.875rem, 1.8vw + 0.4rem, 1.125rem)',
+    fontWeight: 500,
+    lineHeight: 1.4,
+  },
+
+  // Body
+  body1: {
+    fontSize: 'clamp(1rem, 1.8vw + 0.5rem, 1.25rem)',
+    fontWeight: 400,
+    lineHeight: 1.5,
+  },
+  body2: {
+    fontSize: 'clamp(0.875rem, 1.5vw + 0.4rem, 1.125rem)',
+    fontWeight: 400,
+    lineHeight: 1.5,
+  },
+
+  // Caption / small text
+  caption: {
+    fontSize: 'clamp(0.75rem, 1.2vw + 0.35rem, 1rem)',
+    fontWeight: 400,
+    lineHeight: 1.4,
+  },
+};
