+++ src/utils/formatUtcHelper.ts
@@ -1,0 +1,6 @@
+export default function formatUtcHelper(utcDate: string): string {
+  return new Intl.DateTimeFormat('en-US', {
+    year: 'numeric',
+    month: 'short',
+  }).format(new Date(utcDate));
+}
